backend/requirements.txt
asgiref==3.8.1
django==5.1.4
djangorestframework==3.15.2
mysqlclient==2.2.6
pillow==11.0.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2024.2
